Move hapi to legacyHapi add a test
