+++ legacy-hapi.js
